Create SetupInstructions.txt
Since I'm not working on this continuously, it's good to have a place to remind myself of all I did to set things up. 

There are things that happen on the Mac and some on the Onion. Potentially, parts of this could be an environment setup script that could just set things up on each side.